refactor: unify supabase utilities
