+++ app/auth/callback/route.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/utils/supabase/server'
+import { createServerClient as createClient } from '@/lib/supabase/server'
 import { NextResponse } from 'next/server'
 
 export async function GET(request: Request) {
