+++ app/games/page.tsx
@@ -2,7 +2,8 @@
 // ABOUTME: Server component that fetches user's games
 import { Suspense } from 'react'
 import { redirect } from 'next/navigation'
-import { createServerClient, requireAuth } from '@/utils/supabase'
+import { createServerClient } from '@/lib/supabase/server'
+import { requireAuth } from '@/utils/supabase/auth'
 import { ProtectedRoute } from '@/components/auth/protected-route'
 import { LoadingSpinner } from '@/components/ui/loading-spinner'
 import { GamesContent } from './games-content'
