+++ app/questions/[id]/edit/page.tsx
@@ -19,7 +19,7 @@ import { LoadingSpinner } from '@/components/ui/loading-spinner'
 import { ProtectedRoute } from '@/components/auth/protected-route'
 import { useToast } from '@/hooks/use-toast'
 import { errorHandler } from '@/lib/errors/handler'
-import { createClient } from '@/utils/supabase/client'
+import { createClient } from '@/lib/supabase/client'
 import { QuestionSetGridSkeleton } from '@/components/questions/question-set-grid-skeleton'
 import { debounce } from 'lodash'
 import { TagInput } from 'emblor'
