+++ lib/supabase/question-sets.ts
@@ -1,4 +1,4 @@
-import { createClient } from '@/utils/supabase/client'
+import { createClient } from '@/lib/supabase/client'
 import type { TablesInsert } from '@/lib/supabase/database.types'
 
 export interface CreateQuestionSetData {
