+++ middleware.ts
@@ -1,9 +1,9 @@
 import { type NextRequest, NextResponse } from 'next/server'
-import { createClient } from '@/utils/supabase/middleware'
+import { createMiddlewareClient } from '@/lib/supabase/middleware'
 
 export async function middleware(request: NextRequest) {
   try {
-    const { supabase, response } = createClient(request)
+    const { supabase, response } = createMiddlewareClient(request)
 
     // Refresh session if expired - required for Server Components
     const { data: { user }, error } = await supabase.auth.getUser()
